<Android App> 找房页面逻辑（显示所有房源）

+++ App_Android/app/src/main/java/com/kjs/skywalk/app_android/ClassDefine.java
@@ -69,7 +69,7 @@ public class ClassDefine {
         public int Bedrooms;
         public int Livingrooms;
         public int Bathrooms;
-        public int Acreage;
+        public double Acreage;
         public int Rental;
         public int Pricing;
         public int CoverImage;
